added SocialLink js and css files
