+++ client/src/css/components/socialLink.css
@@ -1,0 +1,4 @@
+.SocialLink button{
+ 
+    
+}
