--- client/src/js/components/SigninLink.js
@@ -1,22 +1,0 @@
-import React from 'react'
-import { App } from './App'
-
-
-class SigninControl extends React.Component {
-    constructor(props) {
-        super(props);
-        this.handleLoginClick = this.handleLoginClick.bind(this);
-        this.handleLogoutClick = this.handleLogoutClick.bind(this);
-        this.state = {isLoggedIn: false};
-      }
-    
-      handleLoginClick() {
-        this.setState({isLoggedIn: true});
-      }
-    
-      handleLogoutClick() {
-        this.setState({isLoggedIn: false});
-      }
-}
-
-export { SigninControl }
